Move ImageUploader to flat structure
