+++ packages/client/src/components/form/DocumentUploadfield/SimpleDocumentUploader.tsx
@@ -9,7 +9,7 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-import { ImageUploader } from '@opencrvs/components/lib/forms'
+import { ImageUploader } from '@opencrvs/components/lib/ImageUploader'
 import { ErrorText } from '@opencrvs/components/lib/forms/ErrorText'
 import { DocumentPreview } from '@client/components/form/DocumentUploadfield/DocumentPreview'
 import { IFormFieldValue, IAttachmentValue } from '@client/forms'
