+++ packages/components/src/components/forms/index.ts
@@ -20,9 +20,7 @@ export * from './SubSectionDivider'
 export * from './VerifyingIndicator'
 export * from './ValidIndicator'
 export * from './CheckboxGroup/CheckboxGroup'
-export * from './ImageUploader'
 export * from './PDFViewer/PDFViewer'
 export * from './ErrorMessage'
 export * from './InputField/InputLabel'
-export * from './ProgressBar'
 export * from './FormTabs'
